requirements.txt
python-telegram-bot[job-queue]==20.7
google-generativeai==0.7.2
Pillow==10.4.0
notion-client==2.2.1
python-dotenv==1.0.1
